test: 💍 add casfs tests on top of fs
